Tried to use dependency injection for the screen, in order to decouple the code more, but I think I made it worse
