+++ Tetfuza/TetfuzaBackend/TetfuzaBackend.cs
@@ -3,35 +3,12 @@ using System;
 using FramerateStabilizer;
 using System.Diagnostics;
 using System.Linq;
+using Tetfuza.Interfaces;
 
 namespace Tetfuza
 {
-    public class SystemConsole : IConsole
-    {
-        public ConsoleKey ReadKey()
-        {
-            return System.Console.ReadKey().Key;
-        }
-
-        public void Clear()
-        {
-            Console.Clear();
-        }
-
-        public bool KeyAvailable
-        {
-            get { return Console.KeyAvailable; }
-        }
-    }
-
     public class TetfuzaBackend
     {
-        public const int BOARD_HEIGHT = 22;
-        public const int BOARD_WIDTH = 10;
-        public const char LOCKDOWN_CHAR = '0';
-        public const char MOVING_CHAR = '@';
-        public const char EMPTY_CHAR = ' ';
-        public const char TOPOUT_CHAR = '%';
         public const int MS_PER_FRAME = 17;
 
         private Random _rand = new Random();
@@ -39,11 +16,14 @@ namespace Tetfuza
         private Stopwatch _timer = new Stopwatch();
         private int _frameCount;
         private InputChecker _keyboard;
+        private IDisplay _screen;
         private bool _userInputDown;
+        private TetfuzaBoard Board;
+
 
         public FuzaPiece CurrentPiece { get; private set; }
         public FuzaPiece AfterPiece { get; private set; }
-        public List<List<char>> Board { get; private set; }
+
         public long Score { get; private set; } = 0;
         public int Lines { get; private set; } = 0;
         public int StartLevel { get; private set; } = 0;
@@ -89,70 +69,27 @@ namespace Tetfuza
                 return dropFrameDelay;
             }
         }
-
-        private List<char> EmptyLine
-        {
-            get
-            {
-                return BoardLine(EMPTY_CHAR);
-            }
-        }
         
         /// <summary>
         /// Creates a new instance of a TetfuzaBackend, and initializes an empty board
         /// </summary>
-        public TetfuzaBackend()
+        public TetfuzaBackend(IConsole keyboard, IDisplay screen)
         {
-            Board = BoardInit();
-            _keyboard = new InputChecker(new SystemConsole());
+            Board = new TetfuzaBoard();
+            _keyboard = new InputChecker(keyboard);
+            _screen = screen;
         }
+
         /// <summary>
         /// Starts the game on a specified level (which determines automatic drop speed and point multiplier)
         /// </summary>
         /// <param name="startLevel"></param>
-        public TetfuzaBackend(int startLevel)
+        public TetfuzaBackend(IConsole keyboard, IDisplay screen, int startLevel)
         {
-            Board = BoardInit();
+            Board = new TetfuzaBoard();
             StartLevel = startLevel;
-            _keyboard = new InputChecker(new SystemConsole());
-        }
-        /// <summary>
-        /// Create a new board with {BOARD_HEIGHT} rows, and {BOARD_WIDTH} columns, all filled with ' ' chars
-        /// </summary>
-        /// <returns></returns>
-        private List<List<char>> BoardInit()
-        {
-            var startingBoard = new List<List<char>>();
-            for (int height = 0; height < BOARD_HEIGHT; height++)
-            {
-                startingBoard.Add(EmptyLine);
-            }
-            return startingBoard;
-        }
-
-        private List<char> BoardLine(char fillerChar)
-        {
-            var line = new List<char>();
-            for (int i = 0; i < BOARD_WIDTH; i++)
-            {
-                line.Add(fillerChar);
-            }
-            return line;
-        }
-
-        public override string ToString()
-        {
-            string boardString = "";
-            for (int row = 2; row < Board.Count; row++)
-            {
-                for (int col = 0; col < BOARD_WIDTH; col++)
-                {
-                    boardString += Board[row][col] + " ";
-                }
-                boardString += ")";
-            }
-            
-            return boardString;
+            _keyboard = new InputChecker(keyboard);
+            _screen = screen;
         }
 
         /// <summary>
@@ -174,6 +111,10 @@ namespace Tetfuza
                 // Determine next piece
                 pieceNum = _rand.Next(0, 7);
                 AfterPiece = new FuzaPiece((FuzaType)pieceNum);
+
+                // Draw Board
+                _screen.RedrawFrame();
+                _screen.DrawBoard(Board.ToString());
 
                 bool isLockDown = false;
                 bool inputDown = false;
@@ -206,73 +147,29 @@ namespace Tetfuza
                         isLockDown = !DropPiece();
                     }
 
-                    DrawPiece();
+                    
+                    Board.DrawPiece(CurrentPiece, _pieceCenter);
+
+                    // Draw Screen
+                    _screen.RedrawFrame();
+                    _screen.DrawBoard(Board.ToString());
 
                     StableFrames.Stabilize(MS_PER_FRAME, _timer);
                     _frameCount++;
                 }
                 // Represent locked piece on board with a different character
-                DrawPiece(LOCKDOWN_CHAR);
+                Board.DrawPiece(CurrentPiece, _pieceCenter, TetfuzaBoard.LOCKDOWN_CHAR);
 
                 // Check if lines were cleared and update score
-                int linesCleared = CheckClear();
-                Lines += linesCleared;
-                int scoreMultiplier = 0;
-                switch (linesCleared)
-                {
-                    case 1:
-                        scoreMultiplier = 40;
-                        break;
-                    case 2:
-                        scoreMultiplier = 100;
-                        break;
-                    case 3:
-                        scoreMultiplier = 300;
-                        break;
-                    case 4:
-                        scoreMultiplier = 1200;
-                        break;
-                }
-                Score += scoreMultiplier * (Level + 1);
+                CheckClear();
+                
 
                 // Check if player has topped out (game over)
                 gameOver = CheckTopOut();
             }
             // Plays the game over animation
-            TopOutAnimation();
+            Board.TopOutAnimation(MS_PER_FRAME, _timer, _screen);
             return Score;
-        }
-
-
-
-        /// <summary>
-        /// Takes data from CurrentPiece.Piece (a 2-D List that represents piece orientation), 
-        /// and copies it to the Board, centered around member variable "_pieceCenter"
-        /// </summary>
-        /// <param name="fuzaChar">The character to write to the Board</param>
-        private void DrawPiece(char fuzaChar = MOVING_CHAR)
-        {
-            for (int row = 0; row < FuzaPiece.PIECE_SIZE; row++)
-            {
-                for (int col = 0; col < FuzaPiece.PIECE_SIZE; col++)
-                {
-                    char fuza = CurrentPiece.Piece[row][col];
-                    int colCoord = _pieceCenter.xPos - (FuzaPiece.PIECE_SIZE / 2) + col;
-                    int rowCoord = _pieceCenter.yPos - (FuzaPiece.PIECE_SIZE / 2) + row;
-                    if (colCoord >= 0 && colCoord < BOARD_WIDTH
-                        && rowCoord >= 0 && rowCoord < BOARD_HEIGHT)
-                    {
-                        if (fuza != FuzaPiece.BLANK_CHAR)
-                        {
-                            Board[rowCoord][colCoord] = fuzaChar;
-                        }
-                        else if (fuza == FuzaPiece.BLANK_CHAR && Board[rowCoord][colCoord] != LOCKDOWN_CHAR)
-                        {
-                            Board[rowCoord][colCoord] = EMPTY_CHAR;
-                        }
-                    }
-                }
-            }
         }
 
         /// <summary>
@@ -311,33 +208,7 @@ namespace Tetfuza
         /// <returns></returns>
         private bool CheckTopOut()
         {
-            return Board[2][6] != EMPTY_CHAR;
-        }
-
-        /// <summary>
-        /// Draws the TOPOUTCHAR over the board one line at a time, starting at the top
-        /// </summary>
-        private void TopOutAnimation()
-        {
-            for (int i = 0; i < BOARD_HEIGHT; i++)
-            {
-                StableFrames.Stabilize(MS_PER_FRAME * 5, _timer);
-                //Board.RemoveAt(BOARD_HEIGHT - 1);
-                //Board.Insert(0, BoardLine(TOPOUT_CHAR));
-                Board[i] = BoardLine(TOPOUT_CHAR);
-            }
-
-            var gameO1 = new List<char>();
-            gameO1.AddRange(BoardLine(TOPOUT_CHAR));
-            gameO1.InsertRange(3, "GAME");
-            Board[BOARD_HEIGHT / 2] = gameO1;
-
-            var gameO2 = new List<char>();
-            gameO2.AddRange(BoardLine(TOPOUT_CHAR));
-            gameO2.InsertRange(3, "OVER");
-            Board[BOARD_HEIGHT / 2 + 1] = gameO2;
-
-            StableFrames.Stabilize(MS_PER_FRAME*4, _timer);
+            return Board.Board[2][6] != TetfuzaBoard.EMPTY_CHAR;
         }
 
         /// <summary>
@@ -356,11 +227,11 @@ namespace Tetfuza
                     char fuza = orientation.Piece[row][col];
                     int colCoord = center.xPos - (FuzaPiece.PIECE_SIZE / 2) + col;
                     int rowCoord = center.yPos - (FuzaPiece.PIECE_SIZE / 2) + row;
-                    if (((colCoord < 0 || colCoord >= BOARD_WIDTH
-                        || rowCoord < 0 || rowCoord >= BOARD_HEIGHT)
+                    if (((colCoord < 0 || colCoord >= TetfuzaBoard.BOARD_WIDTH
+                        || rowCoord < 0 || rowCoord >= TetfuzaBoard.BOARD_HEIGHT)
                         && fuza == FuzaPiece.FUZA_CHAR )
                         || (fuza == FuzaPiece.FUZA_CHAR 
-                        && Board[rowCoord][colCoord] == LOCKDOWN_CHAR))
+                        && Board.Board[rowCoord][colCoord] == TetfuzaBoard.LOCKDOWN_CHAR))
                     {
                         isValid = false;
                     }
@@ -376,23 +247,30 @@ namespace Tetfuza
         /// <returns></returns>
         private int CheckClear()
         {
-            int linesCleared = 0;
-            List<int> clearedIndex = new List<int>();
-            for (int i = 0; i < BOARD_HEIGHT; i++)
+            // Have board clear lines and return the amount cleared
+            int linesCleared = Board.ClearLines(MS_PER_FRAME, _timer, _screen);
+
+            // Update total line count
+            Lines += linesCleared;
+
+            // Update total score
+            int scoreMultiplier = 0;
+            switch (linesCleared)
             {
-                if (!(Board[i].Contains(EMPTY_CHAR) || Board[i].Contains(MOVING_CHAR)))
-                {
-                    linesCleared++;
-                    Board[i] = BoardLine('-');
-                    clearedIndex.Add(i);
-                }
+                case 1:
+                    scoreMultiplier = 40;
+                    break;
+                case 2:
+                    scoreMultiplier = 100;
+                    break;
+                case 3:
+                    scoreMultiplier = 300;
+                    break;
+                case 4:
+                    scoreMultiplier = 1200;
+                    break;
             }
-            StableFrames.Stabilize(MS_PER_FRAME * 10, _timer);
-            for (int index = 0; index < linesCleared; index++)
-            {
-                Board.RemoveAt(clearedIndex[index]);
-                Board.Insert(0, EmptyLine);
-            }
+            Score += scoreMultiplier * (Level + 1);
             return linesCleared;
         }
 
@@ -401,8 +279,13 @@ namespace Tetfuza
             bool isValid = CheckMove(new Coordinate(_pieceCenter.xPos, _pieceCenter.yPos + 1), CurrentPiece);
             if (isValid)
                 _pieceCenter.yPos = _pieceCenter.yPos + 1;
-            DrawPiece();
+            Board.DrawPiece(CurrentPiece, _pieceCenter);
             return isValid;
+        }
+
+        public override string ToString()
+        {
+            return Board.ToString();
         }
     }
 }
