+++ Tetfuza/TetfuzaTests/InputCheckerTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Tetfuza;
+using Tetfuza.Interfaces;
 
 namespace TetfuzaTests
 {
